aggregator/requirements.txt
Flask==2.2.2
Werkzeug>=2.2.0,<3.0.0
pandas==1.5.3
numpy==1.22.4
requests==2.28.2
gunicorn==20.1.0
cachetools==5.3.0
PyJWT==2.7.0
